docs: Add letsencrypt docker-compose example
This addition is meant to address the lack of exposure of how to use the linuxserver letsencrypt image with bookstack together even though they work amazingly well with each other.  

I think many would want to have an https example with bookstack and might be turned away by the digging that's required to figure it out.  Setting up nginx and letsencrypt is not so much fun :)